pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kitten_palentir"]

[project]
name = "kitten-palentir"
version = "0.1.0"
description = "Advanced multi-agent OSINT intelligence platform for company reconnaissance using CAMEL-AI, Neo4j, and MCP tools"
readme = "README.md"
requires-python = "==3.12.*"
license = {text = "Apache-2.0"}
authors = [
    {name = "OSINT Team", email = "team@palentir-osint.local"}
]
keywords = ["osint", "reconnaissance", "multi-agent", "camel-ai", "neo4j", "graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Internet",
]

dependencies = [
    # Core Framework
    "camel-ai[all]>=0.2.82",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Web Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # Frontend
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    
    # Database & Storage
    "neo4j>=5.14.0",
    "qdrant-client",
    "psycopg>=3.2",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "redis>=5.0.0",
    
    # Data Processing
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "pydantic-core>=2.10.0",
    
    # LLM & Embeddings
    "openai>=1.3.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "sentence-transformers>=2.2.0",
    
    # Web & API Clients
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "selenium>=4.14.0",
    "playwright>=1.40.0",
    
    # MCP Integration
    "mcp>=0.1.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "typer>=0.9.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    
    # Async & Concurrency
    "anyio>=4.1.0",
    "asyncio-contextmanager>=1.0.0",
    
    # Validation & Serialization
    "marshmallow>=3.20.0",
    "orjson>=3.9.0",
    
    # Monitoring & Logging
    "prometheus-client>=0.18.0",
    "structlog>=23.2.0",
    
    # Security
    "cryptography>=41.0.0",
    "pyjwt>=2.8.0",
    
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    # pytest-httpx
]

[project.optional-dependencies]
dev = [
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "pylint>=3.0.0",
    "mypy>=1.7.0",
    "pytest-xdist>=3.5.0",
    "pytest-timeout>=2.2.0",
    "ipython>=8.17.0",
    "ipdb>=0.13.0",
]

docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=1.25.0",
    "myst-parser>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/palentir-osint/palentir-osint"
Documentation = "https://docs.palentir-osint.local"
Repository = "https://github.com/palentir-osint/palentir-osint.git"
Issues = "https://github.com/palentir-osint/palentir-osint/issues"

[tool.uv]



[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*"]

[tool.mypy]
plugins = ['pydantic.mypy']
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__main__.py",
    "*/setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.pylint.messages_control]
disable = [
    "C0330", "C0326",  # Bad whitespace
    "R0913",           # Too many arguments
    "R0914",           # Too many local variables
    "W0212",           # Protected access
    "C0103",           # Invalid name
]

[tool.pylint.format]
max-line-length = 100

[tool.pylint.design]
max-attributes = 10
max-arguments = 7
